register user

+++ src/utils/cloudinary.js
@@ -1,31 +1,32 @@
-import {v2 as cloudinary} from "cloudinary"
-import { response } from "express";
-import fs from "fs"
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
-cloudinary.config({ 
-        cloud_name: CLOUDINARY_CLOUD_NAME, 
-        api_key: CLOUDINARY_API_KEY, 
-        api_secret: CLOUDINARY_API_SECRET
-    });
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
+});
 
-
-
-const uploadOnCloudinary= async (localFilePath)=>{
+const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!loacalFilePath) return null
-        //upload the file on cloudinary
-        cloudinary.uploader.upload(localFilePath,{
-            resource_type: "auto"
-        })
-        //file has been uploaded successfully
-        console.log("file is uploaded in cloudinary",response.url)
-        return response
+
+        if (!localFilePath) return null;
+
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto",
+        });
+
+        fs.unlinkSync(localFilePath);
+
+        return response;
+
     } catch (error) {
-        fs.unlinkSync(localFilePath)
-        //removes the locally saved temporary file as the upload operation got failed
-        return null
+
+        if (localFilePath) {
+            fs.unlinkSync(localFilePath);
+        }
+        return null;
     }
+};
 
-}
-
-export {uploadOnCloudinary}
+export { uploadOnCloudinary };
